fix cors, moved remove functionality to account endpoint
