+++ backend/src/StamAcasa.IdentityServer/Quickstart/Account/DeleteAccountModel.cs
@@ -1,8 +1,14 @@
 ﻿
+using System.ComponentModel.DataAnnotations;
+
 namespace StamAcasa.IdentityServer.Quickstart.Account
 {
     public class DeleteAccountModel
     {
+        [Required]
+        public string Email { get; set; }
+
+        [Required]
         public string Password { get; set; }
     }
 }
