+++ frontend/src/api/accountApi.js
@@ -2,12 +2,11 @@ import axios from "axios";
 import { Constants } from "../config/constants";
 
 const api = axios.create({
-    baseURL: `${Constants.idpUrl}/api/`
+  baseURL: `${Constants.idpUrl}/account/`
 });
 
 const AccountApi = {
-    deleteAccount: (password) =>api.post("delete", { password })
-    
+  deleteAccount: (email, password) => api.post("delete", { email, password })
 };
 
 export default AccountApi;
